requirements.txt
absl-py==2.0.0
affine==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
alabaster==0.7.13
albumentations==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
bio==1.5.9
biopython==1.81
biothings-client==0.3.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
commonmark==0.9.1
contourpy==1.1.1
cycler==0.12.1
dill==0.3.7
docutils==0.20.1
einops==0.7.0
et-xmlfile==1.1.0
filelock==3.12.4
fiona==1.9.5
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.1
fsspec==2023.9.2
gast==0.5.4
gensim==4.3.2
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gprofiler-official==1.0.0
grpcio==1.59.0
h5py==3.10.0
huggingface-hub==0.17.3
idna==3.4
imagecodecs==2024.1.1
imageio==2.33.1
imagesize==1.4.1
isodate==0.6.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
keras==2.14.0
keras-tuner==1.4.6
kiwisolver==1.4.5
kt-legacy==1.0.5
lazy_loader==0.3
libclang==16.0.6
lightning-utilities==0.10.0
littleutils==0.2.2
llvmlite==0.41.1
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-venn==0.11.9
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.4
mygene==3.2.2
networkx==3.1
numba==0.58.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
outdated==0.2.2
palettable==3.3.3
pandarallel==1.6.5
pandas==2.1.1
pandas-flavor==0.6.0
patsy==0.5.4
Pillow==10.0.1
pingouin==0.5.3
platformdirs==3.11.0
plyfile==1.0.2
pooch==1.7.0
progressbar2==4.3.2
protobuf==4.24.4
PubChemPy==1.0.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pynndescent==0.5.11
pyparsing==3.1.1
pyproj==3.6.1
python-utils==3.8.1
pytorch-lightning==2.1.3
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
qudida==0.0.4
rasterio==1.3.9
rdflib==7.0.0
rdkit==2023.3.3
recommonmark==0.7.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
Rtree==1.1.0
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
selfies==2.1.1
sentencepiece==0.1.99
shapely==2.0.2
slidingwindow==0.0.14
smart-open==6.4.0
snowballstemmer==2.2.0
snuggs==1.4.7
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
statsmodels==0.14.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.12.9
tokenizers==0.14.1
tqdm==4.66.1
transformer-pytorch==0.0.1
transformers==4.34.0
tzdata==2023.3
umap-learn==0.5.5
urllib3==2.0.6
Werkzeug==3.0.0
wrapt==1.14.1
xarray==2023.11.0
xmltodict==0.13.0
yarl==1.9.4
